Added Group Icon to post editor, updating icon admin page.
