+++ zapread.com/Views/Home/Index.cshtml
@@ -85,7 +85,7 @@
                 <span></span>
             </div>
             <div class="social-feed-box-nb">
-                <button id="btnNewPost" onclick="location.href='@Url.Action("NewPost", "Post")'" class="btn btn-primary btn-block"><i class="fa fa-plus"></i> Add Post</button>
+                <button id="btnNewPost" onclick="location.href='@Url.Action("Edit", "Post")'" class="btn btn-primary btn-block"><i class="fa fa-plus"></i> Add Post</button>
             </div>
             <div id="posts">
                 <div class="ibox-content no-padding sk-loading" style="border-style:none">
