+++ zapread.com/package.json
@@ -63,6 +63,7 @@
     "react-bootstrap-table2-paginator": "2.1.2",
     "react-bs-datatable": "2.1.0",
     "react-dom": "^16.13.1",
+    "react-dropzone": "11.0.1",
     "react-quill": "2.0.0-beta.2",
     "sass": "1.26.3",
     "sass-loader": "8.0.2",
